feat/support urls
allow urls as arguments everywhere

allow .md and .html file paths in respect document chunkers
