+++ ovos_document_chunkers/__init__.py
@@ -1,13 +1,7 @@
 from ovos_document_chunkers.text import (SaTSentenceSplitter, SaTParagraphSplitter,
                                          WtPSentenceSplitter, WtPParagraphSplitter,
                                          RegexSentenceSplitter, RegexParagraphSplitter,
-                                         PySBDSentenceSplitter,
-                                         MarkdownSentenceSplitter, MarkdownParagraphSplitter,
-                                         HTMLSentenceSplitter, HTMLParagraphSplitter,
-                                         PDFSentenceSplitter, PDFParagraphSplitter,
-                                         DOCSentenceSplitter, DOCParagraphSplitter,
-                                         DOCxSentenceSplitter, DOCxParagraphSplitter)
-
+                                         PySBDSentenceSplitter)
 
 __all__ = [
     "SaTSentenceSplitter",
@@ -28,4 +22,3 @@ __all__ = [
     "DOCxSentenceSplitter",
     "DOCxParagraphSplitter"
 ]
-
