+++ ovos_document_chunkers/files/__init__.py
@@ -1,0 +1,15 @@
+from ovos_document_chunkers.files.doc import DOCSentenceSplitter, DOCParagraphSplitter
+from ovos_document_chunkers.files.webpages import HTMLSentenceSplitter, HTMLParagraphSplitter
+
+__all__ = [
+    "MarkdownSentenceSplitter",
+    "MarkdownParagraphSplitter",
+    "HTMLSentenceSplitter",
+    "HTMLParagraphSplitter",
+    "PDFSentenceSplitter",
+    "PDFParagraphSplitter",
+    "DOCSentenceSplitter",
+    "DOCParagraphSplitter",
+    "DOCxSentenceSplitter",
+    "DOCxParagraphSplitter"
+]
