+++ ovos_document_chunkers/text/__init__.py
@@ -1,11 +1,5 @@
 from ovos_document_chunkers.text.paragraphs import SaTParagraphSplitter, WtPParagraphSplitter, RegexParagraphSplitter
 from ovos_document_chunkers.text.sentence import PySBDSentenceSplitter, SaTSentenceSplitter, WtPSentenceSplitter, RegexSentenceSplitter
-from ovos_document_chunkers.text.markdown import MarkdownSentenceSplitter, MarkdownParagraphSplitter
-from ovos_document_chunkers.text.webpages import HTMLSentenceSplitter, HTMLParagraphSplitter
-from ovos_document_chunkers.text.pdf import PDFSentenceSplitter, PDFParagraphSplitter
-from ovos_document_chunkers.text.doc import DOCSentenceSplitter, DOCParagraphSplitter
-from ovos_document_chunkers.text.docx import DOCxSentenceSplitter, DOCxParagraphSplitter
-
 
 __all__ = [
     "SaTParagraphSplitter",
@@ -14,13 +8,5 @@ __all__ = [
     "PySBDSentenceSplitter",
     "SaTSentenceSplitter",
     "WtPSentenceSplitter",
-    "RegexSentenceSplitter",
-    "MarkdownSentenceSplitter",
-    "MarkdownParagraphSplitter",
-    "PDFSentenceSplitter",
-    "PDFParagraphSplitter",
-    "DOCSentenceSplitter",
-    "DOCParagraphSplitter",
-    "DOCxSentenceSplitter",
-    "DOCxParagraphSplitter"
+    "RegexSentenceSplitter"
 ]
