+++ setup.py
@@ -53,7 +53,8 @@ setup(
     author_email='jarbasai@mailfence.com',
     license='MIT',
     packages=['ovos_document_chunkers',
-              'ovos_document_chunkers.text'],
+              'ovos_document_chunkers.text',
+              'ovos_document_chunkers.files'],
     zip_safe=True,
     keywords='OVOS openvoiceos plugin',
     install_requires=required("requirements.txt"),
